requirements.txt
pandas
numpy
scikit-learn
nltk
indic-nlp-library
matplotlib
seaborn
streamlit
joblib
